ASoC: qdsp6v2: Update RTC with device pair used in a voice call
Currently, there is no mechanism to intimate rtac driver with rx
and tx device ids used in a voice call which causes inconsistency
in refresh operations of QACT. Also, no device pair information
with QACT leads to uncertainties in the process of tuning cal data
for different devices.
Extract the device information from vocproc cal data in voice driver
and pass it to rtac driver at the start of the call and every device
switch.

CRs-Fixed: 835421
Signed-off-by: Avinash Vaish <avaish@codeaurora.org>
Change-Id: Id664dfa9843db6e4fefbfdad43732060f6a34e67