qcacld: wma: Do not fail for credit check
During suspend, WMA has a check to make sure enough credits are
available and return failure if the check fails but the check is
not necessary as suspend and resume is synchronized between host
and firmware and if firmware is stuck for some reason then host
times out for suspend. With runtime PM feature, this check is
getting hit often and leading to not to allow the system suspend
to take place.

Fixes: b/23059753

Signed-off-by: Kumar Anand <kanand@codeaurora.org>