qcacld-2.0: Make extscan wakelock a timed wakelock
Currently, extscan wakelock is acquired when host receives
extscan cycle start to keep the APPS from going to suspend
and receive all the beacons and probe responses and release the
wakelock after cycle complete event is received.
If the cycle completed event is not received from fw for any reason,
then the wakelock is held for longer time leading to power issues.
Hence convert the extscan wakelock from non-timed to timed.

Fixes: b/23132501

Signed-off-by: Kumar Anand <kanand@codeaurora.org>