mmc: queue: Check for CQE disable state and remove requeuing
1. Add a check in mmc_cmdq_should_pull_reqs to not pull.
Any requests in case if CQE is disabled and card is not suspended.
This means that CQE must be in error state and so mmc_cmdq_thread
should stop issuing requests.

2. Do not requeue the request in case of error in issue path.
Ideally there should not be any error in the issue path
of request. Even if there is an error, we should not
requeue this request as (like in case of DCMD,
blk_end_request with partial bytes of completion will
make sure that request is not finished and still can be
fetched). So requeing the same request which is still in
elevator is illegal.

Change-Id: I0c5e6556d3d881f1a675db4fff4f995de024ddf8
Signed-off-by: Ritesh Harjani <riteshh@codeaurora.org>