msm: mdss: update backlight during unblank for ESD recovery
For a few panels where the Post Init Delay DT property is set,
the backlight is not updated during unblank since it needs to be
done after frame commit. But, for ESD recovery cases, the framework
might not update backlight after a frame commit. Add change to make
sure backlight is updated during ESD recovery in unblank path.

Change-Id: Id6cec15bda0b05d2fe8e158d605d79a8fd423ceb
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>
Signed-off-by: GreyLeshy <andreiskull@gmail.com>