qcacld2.0: Validate the disassoc reason code correctly
When supplicant is sending del station with reason code 2,
driver is comparing this reason code with the disassoc
trigger reasoneLIM_LINK_MONITORING_DISASSOC instead of the
actual disassoc reason code. Due to which the driver assumes
this as an internally generated event and does not send the
disassoc indication to supplicant and the supplicant keeps
retrying with the del station and the connection always
fails.

To address this issue, validate disassoc reason in the
disassoc req with actual disassoc reason codes intead
of disassoc trigger reason code.

CRs-Fixed: 874813
Change-Id: Icd6596bf9b7736ab810978915f1426f2fac409ec
Git-commit: ff7dbace5bdc5b2a87db295ca2ffe560e0537ede
Bug: 25740368
Signed-off-by: Amarnath Hullur Subramanyam <amarnath@codeaurora.org>