qcacld-2.0: Add explicit DSCP 46 (EF) QoS mapping
Currently the default QoS mapping rule maps DSCP 46 (binary 101110) to
a TOS value of 5 (video).  However DSCP 46 is defined by RFC 2598 to
be an Expedited Forwarding per-hop behavior, and this DSCP value is
utilized for VoIP services, therefore override the default mapping for
this DSCP value and instead map it to a TOS value of 6 (voice).

Change-Id: I200258ed7c6772450e3997a05a2d53859cb9bd87
CRs-Fixed: 909991
Bug: 25779250
Signed-off-by: Amarnath Hullur Subramanyam <amarnath@codeaurora.org>