qcacld-2.0: Reset roaming in progress flag once roaming fails
The roaming in progress flag need not be re-set to false when
roaming is successfull, since the session on which it is set
is anyways deleted if roaming is successfull. But if the roaming
fails, there is a chance that either the connection prevails or
a disconnect happens. But in the disconnect handler it is checked
as to if roaming is in progress and will not disconnect if it is
true. So, if roaming fails and falls back to connected state, then
we need to reset the flag which will eventually help disconnect if
needed.

CRs-Fixed: 947880
Change-Id: I69dadc3a5cb2baa983bfa12bbf29fbfbb9b7792a
Git-commit: 6d076cf84a459331fff0b896fa8e05f9c5521386
Bug: 26069494

Signed-off-by: Amarnath Hullur Subramanyam <amarnath@codeaurora.org>