EAS: Fix handling of max frequency
If policy->max is change, the new frequency calculation will be
incorrect. Keep track of the highest maximum since it corresponds
to the max capacity for the CPU.

Still an issue if the cpu is onlined with a reduced policy->max. The
frequency calcs will be incorrect until policy->max is raised to the
maximum value

Change-Id: Ib2cfbc8efe2c56ea0568792745c4b0591a294ae6