diffconfig: enable fstack-protector-strong
enable fstack-protector-strong

Enable protection against stack corruption. (most) Functions with the
possibility of stack corruption  are protected with a canary.

With this change, the angler kernel grows from 11072067 to 11279319
an increase of 1.9%. Performance loss is considered minimal. Security
against stack overflow is greatly improved. [1]

This is the improved version of CONFIG_CC_STACKPROTECTOR=y which is
enabled in shamu's kernel.

[1] https://lwn.net/Articles/584225

Bug: 28837708
Change-Id: Ifad485a543768be5722e937069983c10fcb80ee9
Signed-off-by: Jeff Vander Stoep <jeffv@google.com>
Signed-off-by: Andrey Cherepkov <andreiskull@gmail.com>