FS: Properly free dentry name snapshots in vfs_rename
An error in commit 0a420e78c31b ("BACKPORT: dentry name snapshots")
errantly replaced the fsnotify_oldname_free() with a
take_dentry_name_snapshot() instead of the intended
release_dentry_name_snapshot() as demonstrated in the upstream
commit. This results in taking another snapshot and leaking of the
memory allocated for the snapshot name.

Fixes: 0a420e78c31b ("BACKPORT: dentry name snapshots")
Change-Id: I13356a0cc12fbaf63330d6781ad65ed735e6bc0d
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Signed-off-by: Andrey Cherepkov <andreiskull@gmail.com>