fs: sdfat: Fix frag_ratio formula
When there is no full au in AMAP, if we use the number of non-clean au
divides fsi->used_clusters * CLUS_PER_AU(sb), the frag_ratio is always
smaller than (or equal with) 100%, which is not right.

Actually, frag_ratio should be the the ratio that non-clean aus divides
the number of aus if all used_clusters are contiguous.

Signed-off-by: Noctis Ackerman <noctis.akm@gmail.com>
Signed-off-by: Andrey Cherepkov <andreiskull@gmail.com>